fix: option to filter-out AutoSummary messages
Problem
--------
If I send for summarization messages that contain summaries, we end up correlating across runs. In particular, the LLM tends to over-rely on past AutoSummary runs and under-represent the rest of the messages in today's run.

Solution
--------
Added funcitonality `class SummaryRenderer` to recognize when a message is from AutoSummary, so that we can filter those out before asking an LLM to make new summaries